perf: Optimize token estimation, embeddings, and resource usage
- Probe tiktoken import once at module load instead of on every estimate_tokens() call to avoid repeated import failures in hot path
- Change cosine_similarity dtype from float64 to float32 to reduce memory overhead while maintaining sufficient precision for score thresholds
- Add defensive access patterns for embedder to handle cases where memory backend or embedder may be unavailable
- Validate embedder availability before routing queries and fall back to LLM-only classification when needed
- Fix relative stylesheet paths in HTML frontends (remove leading slash for proper asset resolution)
- Add shared graph-bootstrap.js include to studio log and memory frontends for consistent initialization
- Improve error handling with explicit null checks using getattr() to prevent AttributeError exceptions during embedder access

+++ cognition/reason.py
@@ -90,8 +90,10 @@ def cached_embed_query(embedder, text: str, instruct: str = "") -> np.ndarray:
 
 
 def cosine_similarity(vec_a, vec_b) -> float:
-    a = np.asarray(vec_a, dtype=float)
-    b = np.asarray(vec_b, dtype=float)
+    # float32 throughout: the old float64 default doubled temp memory per
+    # call for precision nothing downstream can use (scores gate thresholds).
+    a = np.asarray(vec_a, dtype=np.float32)
+    b = np.asarray(vec_b, dtype=np.float32)
     na, nb = np.linalg.norm(a), np.linalg.norm(b)
     if na == 0 or nb == 0:
         return 0.0
